% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eisenrep.R
\name{eisenrep}
\alias{eisenrep}
\title{Calculates Reputation as defined by Eisenegger (2005)}
\usage{
eisenrep(data, rating, weigth)
}
\arguments{
\item{data}{}

\item{rating}{}

\item{weigth}{}
}
\value{
Dataframe with the calculated Reputationindex (or Indices)
}
\description{
This Package calculates the Reputation of a Subject following the methods developed by Eisenegger (2005). The Input needs to be a dataframe containing texts (eg. Tweets, Articles, etc.), their rating ("positiv", "negativ", "ambivalent" or "neutral") and optionally a column with numeric values used as weigth (eg. Nr of Retweets, Nr of Recipients, Centrality of Agent, etc.).
}
\examples{
eisenrep(df, rating, centrality)
eisenrep(data=df, rating=rating, centrality=followers_count)

}
